refactor(web): extract the shared setup pieces from onboarding
Reconnecting a computer and rebinding a messaging app are the same work
whether they are met while setting an Agent up for the first time or
reopened later from its settings. Only the frame around them differs — a
numbered flow with a Continue button, or a dialog over a settings page —
so the inside of that frame now lives in `src/setup/` and onboarding
holds it rather than owning it.

What moved: the command block, the QR frame, the countdown, the check
rows and the connect status, together with the check derivation they
render, the words they say, and the styling they carry.

`deriveChecks` and `messagingCliCheck` take the statuses they actually
read rather than onboarding's `ReadinessFacts`, which stays where it is:
it is the shape of a flow in progress, and settings has an Agent that
already exists. Narrowing the input is what lets both feed the same
functions.

The moved styling takes an `ots-` prefix, since `otv2-` names a flow
these rules no longer belong to. Onboarding keeps `otv2-` for what stays
its own.

Behaviour is unchanged.
